requirement.txt
absl-py==2.0.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
babel==2.14.0
beautifulsoup4==4.12.0
bleach==6.1.0
brotlipy==0.7.0
bzip2==1.0.8
ca-certificates==2022.10.11
cachetools==5.3.2
certifi==2022.12.17
cffi==1.15.1
charset-normalizer==2.0.4
comm==0.2.1
einops==0.8.0
ipdb==0.13.13
matplotlib==3.8.2
mpmath==1.3.0
numpy==1.26.3
pandas==2.2.3
pillow==10.2.0
python==3.10.8
pyyaml==6.0.1
scipy==1.14.1
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tinycss2==1.2.1
torch==2.4.1
torchtext==0.4.0
torchvision==0.16.2+cu121
tqdm==4.64.1
